[mastodon-ws] StreamClientをinterfaceに変更
